requirements.txt
Django==3.0.11
djangorestframework==3.12.2
djangorestframework-gis==0.16
django-cors-middleware==1.5.0
python-dateutil==2.8.1
django-filter==2.4.0
psycopg2-binary==2.8.6
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
django-versatileimagefield==2.0.0
django-storages==1.11.1
boto3==1.16.55
django-debug-toolbar==2.2
djangorestframework-csv==2.1.0
django-leaflet==0.27.1
django-storage-swift==1.2.19
environs[django]==7.3.0
sentry-sdk==0.19.5
yapf==0.30.0
